feat: remove parameter-based graceful shutdown function
